requirements.txt
scikit-learn==0.24.2
joblib==1.0.1
gunicorn==20.1.0
Pillow==9.0.1
numpy==1.21.5
tensorflow==2.10.0
keras==2.10.0
ai-core-sdk>=1.12.0
opencv-python==4.6.0.66
transformers==4.23.1
pytorch-lightning==1.7.7
torch==1.5.0